First improvement pass
