+++ LocalMultiplayerAgent/SettingsJsonGenerator/index.css
@@ -1,0 +1,48 @@
+body {
+  background: white;
+  font-family: sans-serif;
+  font-size: 1rem;
+}
+
+main {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 2rem;
+}
+
+div.output {
+  display: flex;
+}
+div.output textarea {
+  box-sizing: border-box;
+  width: 100%;
+  height: 100%;
+}
+
+div.field {
+  margin-top: 1rem;
+}
+div.field:first-of-type {
+  margin-top: 0;
+}
+
+fieldset {
+  border: 0;
+  margin-top: 1rem;
+  border-top: 5px solid skyblue;
+}
+fieldset:first-of-type {
+  margin-top: 0;
+}
+fieldset legend {
+  font-weight: bold;
+  padding: 0.25rem;
+}
+
+span.warning {
+  color: darkred;
+}
+
+span.caution {
+  color: olive;
+}
