--- LocalMultiplayerAgent/SettingsJsonGenerator/main.html
@@ -1,97 +1,0 @@
-<!DOCTYPE html>
-<html>
-<body onload="onInputChange()">
-<script id="myScript" src="scripts.js"></script>
-
-<table><tr><td>
-    
-<p>
-<h1>LocalMultiplayerAgent Json inputs</h1>
-<h4>PlayFab Configuration</h4>
-TitleId: <input onchange="onInputChange()" type="text" id="TitleId" value="">
-<br />
-Region:
-<select onchange="onInputChange()" name="Region" id="Region">
-    <option value="EastUs">EastUs</option>
-    <option value="EastUs2">EastUs2</option>
-    <option value="WestUs">WestUs</option>
-    <option value="WestUs2">WestUs2</option>
-    <option value="NorthCentralUs">NorthCentralUs</option>
-    <option value="SouthCentralUs">SouthCentralUs</option>
-    <option value="CentralUs">CentralUs</option>
-    <option value="WestCentralUs">WestCentralUs</option>
-    <option value="JapanEast">JapanEast</option>
-    <option value="JapanWest">JapanWest</option>
-    <option value="WestEurope">WestEurope</option>
-    <option value="NorthEurope">NorthEurope</option>
-    <option value="BrazilSouth">BrazilSouth</option>
-    <option value="AustrailiaEast">AustrailiaEast</option>
-    <option value="AustrailiaSoutheast">AustrailiaSoutheast</option>
-    <option value="EastAsia">EastAsia</option>
-    <option value="SoutheastAsia">SoutheastAsia</option>
-    <option value="SouthAfricaNorth">SouthAfricaNorth</option>
-    <option value="UAENorth">UAENorth</option>
-    <option value="CentralIndia">CentralIndia</option>
-    <option value="FranceCentral">FranceCentral</option>
-    <option value="KoreaCentral">KoreaCentral</option>
-</select>
-
-<br />
-Initial Players: <input onchange="onInputChange()" type="text" size="100%" id="InitialPlayers" value="Player1, Player2">
-<br />
-<input onchange="onInputChange()" type="checkbox" id="OutputPlaceholders"> Add advanced config json options
-
-<hr />
-<h4>OS/Container Configuration:</h4>
-<select onchange="onInputChange()" name="RunMode" id="RunMode">
-    <option value="WinProcess">Windows Process</option>
-    <option value="WinContainer">Windows Docker Container</option>
-    <option value="LinuxContainer">Linux Docker Container</option>
-</select>
-
-<hr />
-<h4>Asset Configuration (This tool only supports 1 asset currently):</h4>
-Asset Input Zip <span id="LocalFilePathValidate" style="color:red"></span><input onchange="onInputChange()" type="file" size="100%" id="LocalFilePath" value="">
-<br />
-Asset Extract Path: <span id="MountPathValidate" style="color:red"></span><span id="MountPathWarning" style="color:olive"></span><input onchange="onInputChange()" type="text" size="100%" id="MountPath" value="C:\Assets">
-<br />
-
-<hr />
-<h4>General Configuration:</h4>
-Output Log Folder: <input onchange="onInputChange()" type="text" size="100%" id="OutputFolder" value="C:\Assets\WindowsContainerLogOutput">
-<br />
-Start Command: <span id="StartCommandValidate" style="color:red"></span><input onchange="onInputChange()" type="text" size="100%" id="StartCommand" value="MyGameServer.exe -log">
-<br />
-<!-- 1800 ticks is 15 minutes -->
-Transition from Standby->Active after ticks: <input onchange="onInputChange()" type="range" id="NumHeartBeatsForActivateResponse" min="1" max="1800" value="10"><span id="NumHeartBeatsForActivateResponseOutput"></span>
-<br />
-<!-- 172800 ticks is 24 hours -->
-Set Terminate status after ticks: <input onchange="onInputChange()" type="range" id="NumHeartBeatsForTerminateResponse" min="2" max="172800" value="172800"><span id="NumHeartBeatsForTerminateResponseOutput"></span>
-<br />
-Agent Port <input onchange="onInputChange()" type="range" id="AgentListeningPort" min="1024" max="65535" value="56001"><span id="AgentListeningPortOutput"></span>
-
-<hr />
-<h4>Server Port Configuration:</h4>
-GamePort Name <input onchange="onInputChange()" type="text" id="GamePortName" value="UnrealServerGsdkHostPort">
-<br />
-GamePort Protocol
-<select onchange="onInputChange()" name="GamePortProtocol" id="GamePortProtocol">
-    <option value="UDP">UDP</option>
-    <option value="TCP">TCP</option>
-</select>
-
-<br />
-External Port <input onchange="onInputChange()" type="range" id="NodePort" min="1024" max="65535" value="56100"><span id="NodePortOutput"></span>
-<br />
-<!-- TODO: Hide container only input when using WinProcess mode -->
-Container Internal Port <input onchange="onInputChange()" type="range" id="GamePortNumber" min="1024" max="65535" value="30000"><span id="NumberOutput"></span>
-</p>    
-
-</td><td>
-
-<textarea id="outputText" rows="55" cols="150" value="Sample json will populate here"></textarea>
-
-</td></tr><table>
-
-</body>
-</html>
