+++ LocalMultiplayerAgent/SettingsJsonGenerator/package.json
@@ -1,0 +1,29 @@
+{
+	"name": "settingsjsongenerator",
+	"version": "1.0.0",
+	"description": "MultiplayerSettings.json Generator",
+	"main": "index.js",
+	"scripts": {
+		"build": "npm run js && npm run css",
+		"js": "webpack",
+		"css": "sass --no-source-map src/index.scss:index.css"
+	},
+	"author": "",
+	"license": "ISC",
+	"devDependencies": {
+		"sass": "^1.49.0",
+		"ts-loader": "^9.2.6",
+		"typescript": "^4.5.5",
+		"webpack": "^5.67.0",
+		"webpack-cli": "^4.9.2"
+	},
+	"prettier": {
+		"arrowParens": "avoid",
+		"jsxBracketSameLine": true,
+		"printWidth": 120,
+		"semi": true,
+		"tabWidth": 4,
+		"trailingComma": "es5",
+		"useTabs": true
+	}
+}
