+++ LocalMultiplayerAgent/SettingsJsonGenerator/webpack.config.js
@@ -1,0 +1,22 @@
+const path = require("path");
+
+module.exports = {
+    entry: "./src/index.ts",
+    mode: "production",
+    module: {
+        rules: [
+            {
+                test: /\.tsx?$/,
+                use: "ts-loader",
+                exclude: /node-modules/,
+            }
+        ]
+    },
+    resolve: {
+        extensions: [".ts", ".js"]
+    },
+    output: {
+        filename: "index.js",
+        path: path.resolve(__dirname)
+    }
+}
